requirements.txt
alabaster==0.7.10
Babel==2.4.0
click==6.7
colorama==0.3.7
coreapi==2.2.4
coreschema==0.0.4
cycler==0.10.0
decorator==4.0.10
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.10.5
django-allauth==0.31.0
django-bootstrap3==8.0.0
django-ckeditor==5.2.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-fluent-contents==1.1.11
django-fluent-utils==1.3
django-jet==1.0.5
django-parler==1.7
django-polymorphic==1.1
django-registration==2.2
django-tag-parser==3.1
django-template-analyzer==1.6.1
django-wysiwyg==0.8.0
djangorestframework==3.5.4
docutils==0.13.1
future==0.16.0
gunicorn==19.7.1
html5lib==0.999999999
httpie==0.9.9
imagesize==0.7.1
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.5
livereload==2.5.1
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==1.5.3
mkdocs==0.16.2
mysqlclient==1.3.9
numpy==1.11.2
oauthlib==2.0.2
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.9
Pygments==2.1.3
PyMySQL==0.7.9
pyparsing==2.1.10
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2016.10
PyYAML==3.12
requests==2.12.3
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.3
sqlparse==0.2.2
tornado==4.4.2
traitlets==4.3.1
uritemplate==3.0.0
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5
whitenoise==3.3.0
win-unicode-console==0.5
